making exports optional

+++ package/index.js
@@ -1,7 +1,7 @@
 const argv = require("minimist")(process.argv.slice(2));
 const fs = require("fs");
 const { parse } = require("path");
-const { npm_config_name: name } = process.env;
+const { npm_config_name: name, npm_config_x: x } = process.env;
 
 const data = `
 const { Schema, model } = require("mongoose");
@@ -32,7 +32,7 @@ if (!fs.existsSync("../../models")) {
 	fs.mkdirSync("../../models");
 }
 // !
-if (!fs.existsSync("../../models/index.js")) {
+if (!fs.existsSync("../../models/index.js") && x) {
 	fs.writeFileSync("../../models/index.js", ``);
 }
 // !
@@ -44,22 +44,24 @@ fs.writeFile(`../../models/${name}.js`, data, (err) => {
 });
 
 // !
-let testData;
-fs.readFile("../../models/index.js", "utf8", (err, data) => {
-	testData = data;
-	let models = data.slice(data.indexOf("{") + 1, data.indexOf("}"));
-	let newModels = models.concat(`    ${name},`);
-	// let newData = data.replace(/\{{1}\n(.|\n)+\n\}{1}/gm, `{${newModels}}`);
-	let newData = data.replace(/\{(\d|\w|\s|,)+\}/gm, `{${newModels}\n}`);
-	// let newData = data.replace("const", `${newModels}`);
-	fs.writeFileSync(
-		"../../models/index.js",
-		`const ${name} = require("./${name}");\n${newData}`,
-		(err) => {
-			if (err) {
-				console.log(err);
+if (x) {
+	let testData;
+	fs.readFile("../../models/index.js", "utf8", (err, data) => {
+		testData = data;
+		let models = data.slice(data.indexOf("{") + 1, data.indexOf("}"));
+		let newModels = models.concat(`    ${name},`);
+		// let newData = data.replace(/\{{1}\n(.|\n)+\n\}{1}/gm, `{${newModels}}`);
+		let newData = data.replace(/\{(\d|\w|\s|,)+\}/gm, `{${newModels}\n}`);
+		// let newData = data.replace("const", `${newModels}`);
+		fs.writeFileSync(
+			"../../models/index.js",
+			`const ${name} = require("./${name}");\n${newData}`,
+			(err) => {
+				if (err) {
+					console.log(err);
+				}
 			}
-		}
-	);
-});
+		);
+	});
+}
 // !
